+++ mapchete_eo/brdf/__init__.py
@@ -2,17 +2,12 @@ from mapchete_eo.brdf.models import (
     DirectionalModels,
     SensorModel,
     SunModel,
-    apply_brdf_correction,
     get_brdf_param,
 )
-from mapchete_eo.brdf.tools import get_constant_sun_angle, get_sun_angle_array
 
 __all__ = [
     "get_brdf_param",
-    "apply_brdf_correction",
     "SensorModel",
     "SunModel",
     "DirectionalModels",
-    "get_sun_angle_array",
-    "get_constant_sun_angle",
 ]