+++ mapchete_eo/brdf/config.py
@@ -4,9 +4,6 @@ from enum import Enum
 class BRDFModels(str, Enum):
     none = "none"
     HLS = "HLS"
-    HLS_alt = "HLS_alt"
-    sen2agri = "sen2agri"
-    combined = "combined"
     default = "HLS"