+++ mapchete_eo/brdf/models.py
@@ -6,7 +6,7 @@ from numpy.typing import DTypeLike
 import numpy.ma as ma
 from mapchete.io.raster import ReferencedRaster, resample_from_array
 from mapchete.protocols import GridProtocol
-from mapchete.types import Grid, NodataVal
+from mapchete.types import NodataVal
 from rasterio.crs import CRS
 from rasterio.enums import Resampling
 from rasterio.fill import fillnodata
@@ -19,42 +19,41 @@ logger = logging.getLogger(__name__)
 class DirectionalModels:
     def __init__(
         self,
-        angles,
-        f_band_params,
-        sza,
+        angles: Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray],
+        f_band_params: Tuple[float, float, float],
         model=BRDFModels.default,
-        upscale_factor=10000,
-        sun_model_flag=False,
+        brdf_weight: float = 1.0,
+        log10_bands_scale_flag: bool = True,
         dtype: DTypeLike = np.float32,
     ):
         self.angles = angles
         self.f_band_params = f_band_params
-        self.sza = sza
         self.model = BRDFModels(model)
         if self.model == BRDFModels.none:
             raise ValueError("model cannot be BRDFModels.none")
-        self.upscale_factor = upscale_factor
-        self.sun_model_flag = sun_model_flag
+        self.brdf_weight = brdf_weight
+        self.log10_bands_scale_flag = log10_bands_scale_flag
         self.dtype = dtype
 
-    def get_model(self):
-        if self.sun_model_flag is True:
-            return SunModel(
-                self.angles, self.f_band_params, self.sza, self.model
-            ).get_model()
+    def get_sensor_model(self):
+        return SensorModel(
+            self.angles, self.f_band_params, self.model, self.brdf_weight
+        ).get_model()
 
-        else:
-            return SensorModel(
-                self.angles, self.f_band_params, self.sza, self.model
-            ).get_model()
+    def get_sun_model(self):
+        # Keep the SunModel values as is without weighting
+        return SunModel(
+            self.angles, self.f_band_params, self.model, brdf_weight=1.0
+        ).get_model()
 
     def get_band_param(self):
         sensor_model = SensorModel(
-            self.angles, self.f_band_params, self.sza, self.model
+            self.angles, self.f_band_params, self.model, self.brdf_weight
         ).get_model()
 
+        # Keep the SunModel values as is without weighting
         sun_model = SunModel(
-            self.angles, self.f_band_params, self.sza, model=self.model
+            self.angles, self.f_band_params, self.model, brdf_weight=1.0
         ).get_model()
 
         out_param_arr = sun_model / sensor_model
@@ -65,14 +64,25 @@ class DirectionalModels:
         ).astype(self.dtype, copy=False)
 
     def get_corrected_band_reflectance(self, band):
-        return get_corrected_band_reflectance(band, self.get_band_param())
+        return get_corrected_band_reflectance(
+            band, self.get_band_param(), self.log10_bands_scale_flag
+        )
 
 
 class BaseBRDF:
-    # Class with adapted sen2agri model computation
-    # https://userpages.umbc.edu/~martins/PHYS650/maignan%20brdf.pdf
-    # http://www.esa-sen2agri.org/wp-content/uploads/resources/technical-documents/Sen2Agri_DDF_v1.2_ATBDComposite.pdf
-    def __init__(self, angles, f_band_params, sza, model="sen2agri"):
+    # Class with adapted Sentinel-2 Sentinel-Hub Normalization (Also used elsewhere)
+    # Sources:
+    # https://sci-hub.st/https://ieeexplore.ieee.org/document/8899868
+    # https://sci-hub.st/https://ieeexplore.ieee.org/document/841980
+    # https://custom-scripts.sentinel-hub.com/sentinel-2/brdf/
+    # Alt GitHub: https://github.com/maximlamare/s2-normalisation
+    def __init__(
+        self,
+        angles: Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray],
+        f_band_params: Tuple[float, float, float],
+        model: str = "HLS",
+        brdf_weight: float = 1.0,
+    ):
         self.model = model
         # angles should have a form of tuple where:
         # sun_zenith, sun azimuth, view_zenith, view_azimuth
@@ -80,26 +90,20 @@ class BaseBRDF:
         # Convert Degrees to Radians
         # theta is zenith angles
         # phi is azimuth angles
-        self.theta_sun = theta_sun * np.pi / 180
-        self.theta_view = theta_view * np.pi / 180
-        self.phi = (phi_sun - phi_view) * np.pi / 180
-        self.phi = np.where(self.phi < 0, self.phi + 2 * np.pi, self.phi)
+        self.theta_sun = np.deg2rad(theta_sun)
+        self.theta_view = np.deg2rad(theta_view)
+
+        self.phi_sun = np.deg2rad(phi_sun)
+        self.phi_view = np.deg2rad(phi_view)
+        self.phi = np.abs(np.deg2rad(self.phi_sun) - np.deg2rad(self.phi_view))
+        self.phi = np.where(self.phi > np.pi, 2 * np.pi - self.phi, self.phi)
 
         # (Modis based) Parameters for the linear model
         self.f_band_params = f_band_params
-
-        # Constant sun angle
-        if isinstance(sza, (float, int)):
-            self.sza = np.full(self.theta_sun.shape, sza)
-        elif isinstance(sza, np.ndarray):
-            self.sza = sza
-        else:
-            raise TypeError("sza must either be a number or a numpy array")
-
-        self.xsi_0 = np.full(self.theta_sun.shape, 1.5 * np.pi / 180)
+        self.brdf_weight = brdf_weight
 
     # Get delta
-    def delta(self):
+    def delta(self) -> np.ndarray:
         delta = np.sqrt(
             np.power(np.tan(self.theta_sun), 2)
             + np.power(np.tan(self.theta_view), 2)
@@ -108,126 +112,116 @@ class BaseBRDF:
         return delta
 
     # Air Mass
-    def masse(self):
+    def masse(self) -> np.ndarray:
         masse = 1 / np.cos(self.theta_sun) + 1 / np.cos(self.theta_view)
         return masse
 
     # Get xsi
-    def cos_xsi(self):
+    def cos_xsi(self) -> np.ndarray:
         cos_xsi = np.cos(self.theta_sun) * np.cos(self.theta_view) + np.sin(
             self.theta_sun
         ) * np.sin(self.theta_view) * np.cos(self.phi)
         return cos_xsi
 
-    def sin_xsi(self):
+    def sin_xsi(self) -> np.ndarray:
         x = self.cos_xsi()
         sin_xsi = np.sqrt(1 - np.power(x, 2))
         return sin_xsi
 
-    def xsi(self):
+    def xsi(self) -> np.ndarray:
         xsi = np.arccos(self.cos_xsi())
         return xsi
 
     # Function t
-    def cos_t(self):
-        trig = np.tan(self.theta_sun) * np.tan(self.theta_view) * np.sin(self.phi)
-        d = self.delta()
-        # Coeficient for "t" any natural number is good, 1 or 2 are used
-        coef = 1
-        cos_t = coef / self.masse() * np.sqrt(np.power(d, 2) + np.power(trig, 2))
-        cos_t = np.where(cos_t > 1, 1, cos_t)
-        cos_t = np.where(cos_t < -1, -1, cos_t)
+    def cos_t(self) -> np.ndarray:
+        def sec(x):
+            return 1 / np.cos(x)
+
+        # Coeficient for "t" any natural number is good, 2 is used
+        cos_t = (
+            2
+            * np.sqrt(
+                np.power(self.delta(), 2)
+                + np.power(
+                    (
+                        np.tan(self.theta_sun)
+                        * np.tan(self.theta_view)
+                        * np.sin(self.phi)
+                    ),
+                    2,
+                )
+            )
+            / (sec(self.theta_sun) + sec(self.theta_view))
+        )
+
+        cos_t = np.clip(cos_t, -1, 1)
         return cos_t
 
-    def sin_t(self):
-        x = self.cos_t()
-        sin_t = np.sqrt(1 - np.power(x, 2))
-        return sin_t
-
-    def t(self):
-        t = np.arccos(self.cos_t())
+    def t(self) -> np.ndarray:
+        t = np.clip(np.arccos(self.cos_t()), -1, 1)
         return t
 
     # Function FV Ross_Thick, V is for volume scattering (Kernel)
-    def fv(self):
-        fv = (self.masse() / np.pi) * (
-            (self.t() - self.sin_t() * self.cos_t() - np.pi)
-            + (
-                (1 + self.cos_xsi())
-                / (2 * np.cos(self.theta_sun) * np.cos(self.theta_view))
-            )
-        )
+    def fv(self) -> np.ndarray:
+        fv = (
+            ((np.pi / 2 - self.xsi()) * self.cos_xsi() + self.sin_xsi())
+            / (np.cos(self.theta_sun) + np.cos(self.theta_view))
+        ) - (np.pi / 4)
+
         return fv
 
     #  Function FR Li-Sparse, R is for roughness (surface roughness)
-    def fr(self):
-        fr = None
-        a = 1 / (np.cos(self.theta_sun) + np.cos(self.theta_view))
-        if self.model == "sen2agri" or self.model == "combined":
-            # sen2agri formula
-            fr = 4 / (3 * np.pi) * a * (
-                (np.pi / 2 - self.xsi()) * self.cos_xsi() + self.sin_xsi()
-            ) * (1 + 1 / (1 / self.xsi() / self.xsi_0)) - (1 / 3)
-        if "HLS".lower() in self.model.lower():
-            # HLS formula
-            # https://userpages.umbc.edu/~martins/PHYS650/maignan%20brdf.pdf
-            fr = 4 / (3 * np.pi) * a * (
-                (np.pi / 2 - self.xsi()) * self.cos_xsi() + self.sin_xsi()
-            ) - (1 / 3)
+    def fr(self) -> np.ndarray:
+        def sec(x):
+            return 1 / np.cos(x)
+
+        capital_o = (1 / np.pi) * (
+            (self.t() - np.sin(self.t()) * np.cos(self.t()))
+            * (sec(self.theta_sun) + sec(self.theta_view))
+        )
+
+        fr = (
+            capital_o
+            - sec(self.theta_sun)
+            - sec(self.theta_view)
+            + (0.5 * (1 + self.cos_xsi()) * sec(self.theta_sun) * sec(self.theta_view))
+        )
 
         return fr
 
-    def get_model(self):
-        model_value = None
-
-        if self.model == "HLS" or self.model == "combined":
-            # Standard HLS model
-            model_value = (
-                self.f_band_params[0]
-                + self.f_band_params[2] * self.fv()
-                + self.f_band_params[1] * self.fr()
-            )
-
-        if self.model == "sen2agri":
-            # Sen2agri nomalization
-            # http://www.esa-sen2agri.org/wp-content/uploads/resources/technical-documents/Sen2Agri_DDF_v1.2_ATBDComposite.pdf
-            model_value = (
-                1
-                + self.f_band_params[2] * self.fv()
-                + self.f_band_params[1] * self.fr()
-            )
-
-        if self.model == "HLS_alt":
-            # Alt normalization as in 2.2.2 of:
-            # It is a formula that aims to decompose the HLS harmonization
-            # We just use it as an alt formula for BRDF as it is similar to
-            # the sen2agri approach in some ways
-            # https://www.google.com/url?sa=t&rct=j&q=&esrc=s&source=web&cd=&cad=rja&uact=8&ved=2ahUKEwjijoqF6abrAhXJfMAKHanQC28QFjAAegQIAxAB&url=https%3A%2F%2Fwww.mdpi.com%2F2072-4292%2F11%2F6%2F632%2Fpdf&usg=AOvVaw0gCUSpFzULY3ZVICIXvnBa
-            model_value = self.f_band_params[0] * (
-                1
-                + self.f_band_params[2] / self.f_band_params[0] * self.fv()
-                + self.f_band_params[1] / self.f_band_params[0] * self.fr()
-            )
+    def get_model(self) -> np.ndarray:
+        if self.model == "HLS" or self.model == "default":
+            # Standard (HLS) BRDF model
+            if self.brdf_weight != 1.0:
+                model_value = (
+                    self.f_band_params[0]
+                    + self.f_band_params[2] * self.fv() / self.brdf_weight
+                    + self.f_band_params[1] * self.fr() / self.brdf_weight
+                )
+            else:
+                model_value = (
+                    self.f_band_params[0]
+                    + self.f_band_params[2] * self.fv()
+                    + self.f_band_params[1] * self.fr()
+                )
         return model_value
 
 
 class SensorModel(BaseBRDF):
-    def __init__(self, angles, f_band_params, sza, model="sen2agri"):
-        super().__init__(angles, f_band_params, sza, model)
+    def __init__(self, angles, f_band_params, model="HLS", brdf_weight=1.0):
+        super().__init__(angles, f_band_params, model, brdf_weight)
 
 
 class SunModel(BaseBRDF):
-    def __init__(self, angles, f_band_params, sza, model="sen2agri"):
-        super().__init__(angles, f_band_params, sza, model)
+    def __init__(self, angles, f_band_params, model="HLS", brdf_weight=1.0):
+        super().__init__(angles, f_band_params, model, brdf_weight)
         self.theta_view = np.zeros(self.theta_sun.shape)
-        self.theta_sun = self.sza
-        self.phi = np.zeros(self.theta_sun.shape)
 
 
 def get_corrected_band_reflectance(
     band: ma.MaskedArray,
     correction: np.ndarray,
-    correction_weight: float = 1.0,
+    log10_bands_scale_flag: bool = True,
     nodata: NodataVal = 0,
 ) -> ma.MaskedArray:
     """
@@ -254,11 +248,20 @@ def get_corrected_band_reflectance(
             if isinstance(band, ma.MaskedArray)
             else np.where(band == nodata, True, False)
         )
-        if correction_weight != 1.0:
-            # a correction_weight value of >1 should increase the correction, whereas a
-            # value <1 should decrease the correction
-            correction = 1 - (1 - correction) * correction_weight
-        corrected = (band * correction).astype(band.dtype, copy=False)
+
+        if log10_bands_scale_flag:
+            # # Apply BRDF correction to log10 scaled Sentinel-2 data
+            corrected = (
+                np.log10(band.astype(np.float32, copy=False), where=band > 0)
+                * correction
+            ).astype(np.float32, copy=False)
+            # Revert the log to linear
+            corrected = (np.power(10, corrected)).astype(np.float32, copy=False)
+        else:
+            corrected = (band.astype(np.float32, copy=False) * correction).astype(
+                np.float32, copy=False
+            )
+
         if nodata == 0:
             return ma.masked_array(
                 data=np.where(mask, 0, np.clip(corrected, 1, np.iinfo(band.dtype).max)),
@@ -278,13 +281,15 @@ def get_brdf_param(
     detector_footprints: ReferencedRaster,
     viewing_zenith_per_detector: Dict[int, ReferencedRaster],
     viewing_azimuth_per_detector: Dict[int, ReferencedRaster],
-    sun_zenith_angle: float,
     f_band_params: Tuple[float, float, float],
     model: BRDFModels = BRDFModels.default,
+    brdf_weight: float = 1.0,
+    log10_bands_scale_flag: bool = True,
     smoothing_iterations: int = 10,
     dtype: DTypeLike = np.float32,
 ) -> ma.MaskedArray:
     """
+    TODO: This function should be under Sentinel-2 platform as it uses its metadata (detectors and angles)
     Return BRDF parameters.
     """
     # create output array
@@ -310,7 +315,6 @@ def get_brdf_param(
         for detector_id in np.unique(resampled_detector_footprints)
         if detector_id != 0
     ]
-
     # iterate through detector footprints and calculate BRDF for each one
     for detector_id in detector_ids:
         logger.debug("run on detector %s", detector_id)
@@ -343,8 +347,8 @@ def get_brdf_param(
                 viewing_azimuth_per_detector[detector_id].data,
             ),
             f_band_params=f_band_params,
-            sza=sun_zenith_angle,
             model=model,
+            brdf_weight=brdf_weight,
         ).get_band_param()
 
         # interpolate missing nodata edges and return BRDF difference model
@@ -367,81 +371,3 @@ def get_brdf_param(
         model_params.mask[detector_mask] = detector_brdf.mask[detector_mask]
 
     return model_params
-
-
-def apply_brdf_correction(
-    band=None,
-    band_crs=None,
-    band_transform=None,
-    product_crs=None,
-    sun_azimuth_angle_array=None,
-    sun_zenith_angle_array=None,
-    detector_footprints=None,
-    viewing_zenith=None,
-    viewing_azimuth=None,
-    sun_zenith_angle=None,
-    f_band_params=None,
-    model=BRDFModels.default,
-    smoothing_iterations=10,
-):
-    """
-    Return BRDF corrected band reflectance.
-
-    Parameters
-    ----------
-    band : np.ndarray
-        Band reflectance.
-    band_crs : str
-        CRS of band.
-    band_transform : Affine
-        Band Affine object.
-    product_crs : str
-        CRS of product.
-    sun_azimuth_angle_array : dict
-        Dictionary of Azimuth angle grids.
-    sun_zenith_angle_array : dict
-        Dictionary of Zenith angle grids.
-    detector_footprints : dict
-        Detector footprints by detector ID.
-    viewing_zenith : dict
-        Dictionary of Zenith incidence angles by detector ID.
-    viewing_azimuth : dict
-        Dictionary of Zenith incidence angles by detector ID.
-    sun_zenith_angle : float
-        Constant sun angle for product.
-    model : str
-        BRDF model to run.
-    smoothing_iterations : int
-        Smoothness of interpolated angle grids
-    """
-    if not isinstance(band, ma.MaskedArray):  # pragma: no cover
-        raise TypeError("input band must be a masked array")
-    for param, name in [
-        (band_transform, "band_transform"),
-        (product_crs, "product_crs"),
-        (sun_azimuth_angle_array, "sun_azimuth_angle"),
-        (sun_zenith_angle_array, "sun_zenith_angle"),
-        (detector_footprints, "detector_footprints"),
-        (viewing_zenith, "viewing_zenith"),
-        (viewing_azimuth, "viewing_azimuth"),
-        (sun_zenith_angle, "sun_zenith_angle"),
-        (f_band_params, "f_band_params"),
-    ]:
-        if param is None:  # pragma: no cover
-            raise ValueError(f"{name} must be provided")
-    return get_corrected_band_reflectance(
-        band=band,
-        correction=get_brdf_param(
-            grid=Grid(band_transform, band.shape[0], band.shape[1], crs=band_crs),
-            product_crs=product_crs,
-            sun_azimuth_angle_array=sun_azimuth_angle_array,
-            sun_zenith_angle_array=sun_zenith_angle_array,
-            detector_footprints=detector_footprints,
-            viewing_zenith_per_detector=viewing_zenith,
-            viewing_azimuth_per_detector=viewing_azimuth,
-            sun_zenith_angle=sun_zenith_angle,
-            f_band_params=f_band_params,
-            model=model,
-            smoothing_iterations=smoothing_iterations,
-        ),
-    )