+++ CHANGELOG.rst
@@ -2,6 +2,26 @@
 Changelog
 #########
 
+2024.10.3 - 2024-10-22
+----------------------
+
+* cli
+  
+  * Add `--out_dtype` option to `s2-rgb` CLI operation for debuging
+  * Add `--brdf-log10` flag to `s2-rgb` CLI operation for debugging
+
+* core
+  
+  * Added `_apply_sentinel2_bandpass_adjustment` to `read_np_array` in `platforms.sentinel2.product.S2Product` class, toggle with: `apply_sentinel2_bandpass_adjustment` bool flag
+  * cleanup and update the `brdf` function chain, add some typing
+  * the `brdf` now uses only single model based on following sources:
+
+    * https://sci-hub.st/https://ieeexplore.ieee.org/document/8899868
+    * https://sci-hub.st/https://ieeexplore.ieee.org/document/841980
+    * https://custom-scripts.sentinel-hub.com/sentinel-2/brdf/#
+
+  * added with flag into BRDFModelConfig as `log10_bands_scale_flag` for: `brdf` original band scaling is now converting the bands to `log10` and applying the `brdf` correction on top of the `log10` converted data
+  
 
 2024.10.2 - 2024-10-21
 ----------------------
@@ -14,6 +34,7 @@ Changelog
 ----------------------
 
 * core
+
   * Sentinel-2: make sure pydantic can parse scene classification names from configuration